Unbound.Generics.LocallyNameless.Fresh

Description

Global freshness monad.

Documentation

class Monad m => Fresh m where

The Fresh type class governs monads which can generate new globally unique Names based on a given Name.

Minimal complete definition

fresh

Methods

fresh :: Name a -> m (Name a)

Generate a new globally unique name based on the given one.

newtype FreshMT m a

The FreshM monad transformer. Keeps track of the lowest index still globally unused, and increments the index every time it is asked for a fresh name.

Constructors

FreshMT

Fields

unFreshMT :: StateT Integer m a

runFreshMT :: Monad m => FreshMT m a -> m a

Run a FreshMT computation (with the global index starting at zero).

contFreshMT :: Monad m => FreshMT m a -> Integer -> m a

Run a FreshMT computation given a starting index for fresh name generation.

type FreshM = FreshMT Identity

A convenient monad which is an instance of Fresh. It keeps track of a global index used for generating fresh names, which is incremented every time fresh is called.

runFreshM :: FreshM a -> a

Run a FreshM computation (with the global index starting at zero).

contFreshM :: FreshM a -> Integer -> a

Run a FreshM computation given a starting index.