Unbound.Generics.LocallyNameless.LFresh

Description

Local freshness monad.

The LFresh class

class Monad m => LFresh m where

This is the class of monads that support freshness in an (implicit) local scope. Generated names are fresh for the current local scope, not necessarily globally fresh.

Minimal complete definition

lfresh, avoid, getAvoids

Methods

lfresh :: Typeable a => Name a -> m (Name a)

Pick a new name that is fresh for the current (implicit) scope.

avoid :: [AnyName] -> m a -> m a

Avoid the given names when freshening in the subcomputation, that is, add the given names to the in-scope set.

getAvoids :: m (Set AnyName)

Get the set of names currently being avoided.

type LFreshM = LFreshMT Identity

A convenient monad which is an instance of LFresh. It keeps track of a set of names to avoid, and when asked for a fresh one will choose the first unused numerical name.

runLFreshM :: LFreshM a -> a

Run a LFreshM computation in an empty context.

contLFreshM :: LFreshM a -> Set AnyName -> a

Run a LFreshM computation given a set of names to avoid.

newtype LFreshMT m a

The LFresh monad transformer. Keeps track of a set of names to avoid, and when asked for a fresh one will choose the first numeric prefix of the given name which is currently unused.

Constructors

LFreshMT

Fields

unLFreshMT :: ReaderT (Set AnyName) m a

runLFreshMT :: LFreshMT m a -> m a

Run an LFreshMT computation in an empty context.

contLFreshMT :: LFreshMT m a -> Set AnyName -> m a

Run an LFreshMT computation given a set of names to avoid.