Unbound.Generics.LocallyNameless.Shift

Description

The pattern Shift e shifts the scope of the embedded term in e one level outwards.

Documentation

newtype Shift e

The type Shift e is an embedding pattern that shifts the scope of the free variables of the embedded term Embedded e up by one level.

Constructors

Shift e