Unbound.Generics.PermM

Description

A slow, but hopefully correct implementation of permutations.

Documentation

newtype Perm a

A permutation is a bijective function from names to names which is the identity on all but a finite set of names. They form the basis for nominal approaches to binding, but can also be useful in general.

Constructors

Perm (Map a a)

permValid :: Ord a => Perm a -> Bool

permValid p returns True iff the perumation is valid: if each value in the range of the permutation is also a key.

single :: Ord a => a -> a -> Perm a

Create a permutation which swaps two elements.

compose :: Ord a => Perm a -> Perm a -> Perm a

Compose two permutations. The right-hand permutation will be applied first.

apply :: Ord a => Perm a -> a -> a

Apply a permutation to an element of the domain.

support :: Ord a => Perm a -> [a]

The support of a permutation is the set of elements which are not fixed.

isid :: Ord a => Perm a -> Bool

Is this the identity permutation?

join :: Ord a => Perm a -> Perm a -> Maybe (Perm a)

Join two permutations by taking the union of their relation graphs. Fail if they are inconsistent, i.e. map the same element to two different elements.

empty :: Perm a

The empty (identity) permutation.

restrict :: Ord a => Perm a -> [a] -> Perm a

Restrict a permutation to a certain domain.

mkPerm :: Ord a => [a] -> [a] -> Maybe (Perm a)

mkPerm l1 l2 creates a permutation that sends l1 to l2. Fail if there is no such permutation, either because the lists have different lengths or because they are inconsistent (which can only happen if l1 or l2 have repeated elements).